# src/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "ApiRestful-NodeJS",
    "description": "Documentation automatically generated by the <b>swagger.autogen</b> module."
  },
  "host": "localhost:8000",
  "basePath": "/",
  "tags": [
    {
      "name": "get-user",
      "description": "Endpoints"
    }
  ],
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {},
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/user/{userId}": {
      "get": {
        "tags": [
          "get-user"
        ],
        "description": "Endpoint to find a user.",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/user/register": {
      "post": {
        "tags": [
          "create-user"
        ],
        "description": "Endpoint to add a user.",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "User information.",
            "required": true,
            "type": "object",
            "schema": {
              "$ref": "#/definitions/AddUser"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "produces": [
          "application/xml"
        ]
      }
    },
    "/login": {
      "post": {
        "tags": [
          "login"
        ],
        "description": "Endpoint to login.",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "User information.",
            "required": true,
            "type": "object",
            "schema": {
              "$ref": "#/definitions/login"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  },
  "definitions": {
    "user": {
      "type": "object",
      "properties": {
        "nome": {
          "type": "string",
          "example": "Alex Rossi"
        },
        "email": {
          "type": "string",
          "example": "alex@email.com"
        },
        "senha": {
          "type": "string",
          "example": "senha"
        },
        "telefones": {
          "type": "object",
          "properties": {
            "numero": {
              "type": "number",
              "example": 123456789
            },
            "ddd": {
              "type": "number",
              "example": 19
            }
          }
        }
      },
      "xml": {
        "name": "user"
      }
    }
  },
  "AddUser": {
    "$nome": "Alex Rossi",
    "$email": "alex@email.com",
    "$senha": "senha",
    "$telefones": {
      "$numero": 123456789,
      "$ddd": 19
    }
  },
  "login": {
    "$email": "alex@email.com",
    "$senha": "senha"
  }
}